Do the right thing wrt indent guides
To get indent guides, spaces are converted to tabs when loading a file
and tabs are converted back to spaces when writing the file.

This wreaks havoc in files where the distinction TAB vs. space is
important, like Makefiles.

The conversion is moved to a function instead of doing everything in the
AutoCommands.
This function first checks whether doing TAB<->space conversion is ok,
before actually performing the operation.