Ignore changes to self-modifying vim scripts
Some vim scripts extract their documentation on the first invocation and
as a consequence add untracked content to the submodule.

As if that isn't enough, the extracted documentation is then removed
from the script itself, resulting in modified content in the submodule.

Since these things cause git status clutter, we just ignore changes in
the affected submodule trees.